pyproject.toml
[project]
name = "llmq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fal-client>=0.5.6",
    "httpx>=0.27.2",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot2[fastapi]>=2.4.0",
    "openai>=1.54.3",
    "pillow>=11.0.0",
    "tomli>=2.0.2",
]
